Updated Events' unit test suite to match the unit-test branch update
