--- unit_tests/events_tests/events_base.py
@@ -1,21 +1,0 @@
-import tinker
-import unittest
-
-
-class EventsBaseTestCase(unittest.TestCase):
-
-    def setUp(self):
-        self.app = tinker.app.test_client()
-
-    def send_post(self, url, form_contents):
-        return self.app.post(url, data=form_contents, follow_redirects=True)
-
-    def send_get(self, url):
-        return self.app.get(url, follow_redirects=True)
-
-    def tearDown(self):
-        pass
-
-if __name__ == "__main__":
-    testsuite = unittest.TestLoader().discover('.')
-    unittest.TextTestRunner(verbosity=1).run(testsuite)
