+++ unit_tests/events_tests/test_edit_event_page.py
@@ -1,4 +1,4 @@
-from events_base import EventsBaseTestCase
+from . import EventsBaseTestCase
 
 
 class EditEventPageTestCase(EventsBaseTestCase):
