+++ unit_tests/events_tests/test_event_in_workflow.py
@@ -1,4 +1,4 @@
-from events_base import EventsBaseTestCase
+from . import EventsBaseTestCase
 
 class EventInWorkflowTestCase(EventsBaseTestCase):
     #######################
