+++ unit_tests/events_tests/test_submit_edit_form.py
@@ -1,4 +1,4 @@
-from events_base import EventsBaseTestCase
+from . import EventsBaseTestCase
 
 
 class SudmitEditFormTestCase(EventsBaseTestCase):
